Moved all of the timetabler (dof9, ptf9) commands into a separate package, unsync_timetabler.
