+++ unsync_timetabler/unsync_timetabler/dof9/__init__.py
@@ -1,0 +1,6 @@
+unsync_prefix = 'timetabler_dof9'
+
+unsync_commands = [
+    'dates_import.dof9_import_dates',
+    'emergency_teacher_import.dof9_emergency_teacher_import'
+]
