+++ unsync_timetabler/unsync_timetabler/ptf9/__init__.py
@@ -1,0 +1,14 @@
+unsync_prefix = 'timetabler_ptf9'
+
+unsync_commands = [
+    'course_import.ptf9_course_import',
+    'day_import.day_import',
+    'period_import.period_import',
+    'room_import.room_import',
+    'staff_enrollment_import.ptf9_staff_enrollment_import',
+    'staff_import.ptf9_staff_import',
+    'student_enrollment_import.ptf9_student_enrollment_import',
+    'student_import.ptf9_student_import',
+    'student_timetable_import.student_timetable_import',
+    'teacher_timetable_import.teacher_timetable_import',
+]
